fix: Convert num to double in price interval calculation
- Fixed type error in asset_detail_screen.dart line 1009
- Added .toDouble() conversion for math.pow results
- All magnitude calculations now properly convert num to double
- Resolves: "A value of type 'num' can't be assigned to a variable of type 'double'"

The chart axis interval calculation now handles Dart's strict type system correctly.

+++ lib/screens/market/asset_detail_screen.dart
@@ -1006,13 +1006,13 @@ class _AssetDetailScreenState extends State<AssetDetailScreen>
     
     double niceInterval;
     if (normalized <= 1) {
-      niceInterval = 1 * magnitude;
+      niceInterval = (1 * magnitude).toDouble();
     } else if (normalized <= 2) {
-      niceInterval = 2 * magnitude;
+      niceInterval = (2 * magnitude).toDouble();
     } else if (normalized <= 5) {
-      niceInterval = 5 * magnitude;
+      niceInterval = (5 * magnitude).toDouble();
     } else {
-      niceInterval = 10 * magnitude;
+      niceInterval = (10 * magnitude).toDouble();
     }
     
     return niceInterval.toDouble();
